refactor: transformers tests
