+++ vite.config.js
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    include: ['test/*test.js'],
+    include: ['test/**/*test.js'],
     testTimeout: 10000,
   },
 })
